+++ mmk.ts
@@ -2,5 +2,5 @@
 process.env["COMMAND"] = "mmk";
 import { YELLOW, NORMAL_COLOR } from "./prompt";
 process.env["UPDATE_MESSAGE"] = `to update run the command:
-${YELLOW}npm update -g @merrymake/cli${NORMAL_COLOR}`;
+${YELLOW}npm install --global @merrymake/cli@latest${NORMAL_COLOR}`;
 import "./index";